Fixed bug: user enters offset manually
If a user entered the offset manually and it was too high the user would
see that no stylists were found for that city. I now check to see that
if no users were found in the query, but the metadata says otherwise, I
repeat the query with offset=0 instead.